requirements.txt
alembic==1.8.1
black==22.10.0
fastapi==0.81.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
pytest==7.1.3
pytest-cov==4.0.0
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.40
uvicorn==0.18.3
isort==5.12.0
PyJWT==2.3.0
python-multipart==0.0.5
